pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cv-creator"
version = "0.1.0"
description = "Multi-agent CLI application to optimize CVs based on job descriptions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.3",
    "openai>=1.50.0",
    "pdfplumber>=0.11.0",
    "weasyprint>=62.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
cv-creator = "cv_creator.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_creator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
